Fix #48 class and :class at component include
