+++ tests/fixtures/vue-bind/bindings-src.twig
@@ -1,3 +1,3 @@
 <div class="{{class|default('')}}">
-  <img src="{{isBool ? varA : varB}}">
+  <img src="{{ isBool ? varA : varB }}">
 </div>
