+++ tests/fixtures/vue-bind/bindings-with-encoded-attributes.twig
@@ -1,12 +1,12 @@
 <div class="{{class|default('')}}">
-  <a href="{{isBool ? varA.url : varB.url}}">
-    <img src="{{isBool ? varA.src : varB.src}}" alt="{{isBool ? varA.alt : varB.alt}}">
+  <a href="{{ isBool ? varA.url : varB.url }}">
+    <img src="{{ isBool ? varA.src : varB.src }}" alt="{{ isBool ? varA.alt : varB.alt }}">
   </a>
-  <form method="get" action="{{isBool ? varA.url : varB.url}}">
+  <form method="get" action="{{ isBool ? varA.url : varB.url }}">
     <label>
-      <input type="text" name="{{isBool ? varA.name : varB.name}}">
+      <input type="text" name="{{ isBool ? varA.name : varB.name }}">
     </label>
     <input type="submit">
   </form>
-  <a name="{{isBool ? varA.anchor : varB.anchor}}"></a>
+  <a name="{{ isBool ? varA.anchor : varB.anchor }}"></a>
 </div>
