+++ tests/fixtures/vue-bind/bindings-with-filter.twig
@@ -1,6 +1,6 @@
 <div class="{{class|default('')}}">
   <div>
-    <img src="{{imageSrc|myFilter}}">
-    <a title="Title" href="{{imageSrc|myFilter}}">Link</a>
+    <img src="{{ imageSrc|myFilter }}">
+    <a title="Title" href="{{ imageSrc|myFilter }}">Link</a>
   </div>
 </div>
